Extract types into aptible-ai/index, useDashboard hook
